chore(refactor): move query chunking and synthesis code into common crate
