+++ Cargo.toml
@@ -31,7 +31,7 @@ members = [
     "plugins/identity",
     "plugins/linguist",
     "plugins/review",
-    "plugins/typo"]
+    "plugins/typo", "hipcheck-common"]
 
 # Make sure Hipcheck is run with `cargo run`.
 #
