+++ hipcheck-common/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "hipcheck-common"
+description = "Common functionality for the Hipcheck gRPC protocol"
+repository = "https://github.com/mitre/hipcheck"
+version = "0.1.0"
+license = "Apache-2.0"
+edition = "2021"
+publish = false
+
+[dependencies]
+anyhow = "1.0.93"
+log = "0.4.22"
+prost = "0.13.3"
+serde_json = "1.0.133"
+thiserror = "2.0.3"
+tonic = "0.12.3"
+
+[build-dependencies]
+anyhow = "1.0.93"
+pathbuf = "1.0.0"
+tonic-build = "0.12.3"
