+++ hipcheck/build.rs
@@ -6,7 +6,6 @@ use anyhow::Result;
 use pathbuf::pathbuf;
 use schemars::schema_for;
 use std::{fs, path::Path};
-use tonic_build::compile_protos;
 
 fn generate_schemars_for_target_types(out_dir: &Path) -> Result<()> {
 	let out_schemars = vec![("target", schema_for!(Target))];
@@ -18,11 +17,6 @@ fn generate_schemars_for_target_types(out_dir: &Path) -> Result<()> {
 }
 
 fn main() -> Result<()> {
-	// Compile the Hipcheck gRPC protocol spec to an .rs file
-	let root = env!("CARGO_MANIFEST_DIR");
-	let path = pathbuf![root, "proto", "hipcheck", "v1", "hipcheck.proto"];
-	compile_protos(path)?;
-
 	generate_schemars_for_target_types(Path::new("../sdk/schema"))?;
 
 	// Make the target available as a compile-time env var for plugin arch
