+++ sdk/rust/Cargo.toml
@@ -22,11 +22,7 @@ hipcheck-sdk-macros = { path = "../../hipcheck-sdk-macros", version = "0.1.0", o
 typify-macro = "0.2.0"
 url = { version = "2.5.4", features = ["serde"] }
 log = "0.4.22"
-
-
-[build-dependencies]
-anyhow = "1.0.91"
-tonic-build = "0.12.3"
+hipcheck-common = { version = "0.1.0", path = "../../hipcheck-common" }
 
 [features]
 macros = ["dep:hipcheck-sdk-macros"]
