+++ sdk/rust/src/error.rs
@@ -1,6 +1,8 @@
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::proto::{ConfigurationStatus, InitiateQueryProtocolResponse, SetConfigurationResponse};
+use hipcheck_common::proto::{
+	ConfigurationStatus, InitiateQueryProtocolResponse, SetConfigurationResponse,
+};
 use std::{convert::Infallible, error::Error as StdError, ops::Not, result::Result as StdResult};
 use tokio::sync::mpsc::error::SendError as TokioMpscSendError;
 use tonic::Status as TonicStatus;
@@ -62,6 +64,20 @@ pub enum Error {
 
 	#[error(transparent)]
 	Unspecified { source: DynError },
+}
+
+impl From<hipcheck_common::error::Error> for Error {
+	fn from(value: hipcheck_common::error::Error) -> Self {
+		use hipcheck_common::error::Error::*;
+		match value {
+			UnspecifiedQueryState => Error::UnspecifiedQueryState,
+			UnexpectedReplyInProgress => Error::UnexpectedReplyInProgress,
+			ReceivedSubmitWhenExpectingReplyChunk => Error::ReceivedSubmitWhenExpectingReplyChunk,
+			MoreAfterQueryComplete { id } => Error::MoreAfterQueryComplete { id },
+			InvalidJsonInQueryKey(s) => Error::InvalidJsonInQueryKey(s),
+			InvalidJsonInQueryOutput(s) => Error::InvalidJsonInQueryOutput(s),
+		}
+	}
 }
 
 impl From<anyhow::Error> for Error {
